add insertUser2Db function

+++ test/testDao.py
@@ -6,34 +6,55 @@ ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
 
 test_user_db = "testUserDb"
 
-os.remove(ROOT_DIR+'/../db/{}.db'.format(test_user_db))
+if os.path.exists(ROOT_DIR+'/../db/{}.db'.format(test_user_db)):
+    os.remove(ROOT_DIR+'/../db/{}.db'.format(test_user_db))
 
 print("Checking creating test user db...")
-assert(dao.create("CREATE TABLE IF NOT EXISTS user (id int, name varchar(64))", test_user_db) == True)
-assert(len(dao.select("SELECT * FROM user", test_user_db, False)) == 0)
+assert(dao.create("CREATE TABLE IF NOT EXISTS test(id int, name varchar(64))", test_user_db) == True)
+assert(len(dao.select("SELECT * FROM test", test_user_db, False)) == 0)
 
 print("Checking insert test user db...")
-assert(dao.insert("INSERT INTO user(id, name) VALUES(?,?)", [123, "taro"], test_user_db) == True)
-assert(len(dao.select("SELECT * FROM user", test_user_db, False)) == 1)
-
+assert(dao.insert("INSERT INTO test(id, name) VALUES(?,?)", [123, "taro"], test_user_db) == True)
+assert(len(dao.select("SELECT * FROM test", test_user_db, False)) == 1)
 
 print("Checking update test user db...")
-assert(dao.update("UPDATE user SET name=? WHERE id=?", ["goro", 123], test_user_db) == True)
-assert(len(dao.select("SELECT * FROM user", test_user_db, False)) == 1)
-assert(dao.select("SELECT * FROM user", test_user_db, False)[0][1] == "goro")
+assert(dao.update("UPDATE test SET name=? WHERE id=?", ["goro", 123], test_user_db) == True)
+assert(len(dao.select("SELECT * FROM test", test_user_db, False)) == 1)
+assert(dao.select("SELECT * FROM test", test_user_db, False)[0][1] == "goro")
 
+print("Checking delete a column test test db...")
+assert(dao.delete("DELETE FROM test WHERE id=?", [123], test_user_db) == True)
+assert(len(dao.select("SELECT * FROM test", test_user_db, False)) == 0)
 
-print("Checking delete a column test user db...")
-assert(dao.delete("DELETE FROM user WHERE id=?", [123], test_user_db) == True)
-assert(len(dao.select("SELECT * FROM user", test_user_db, False)) == 0)
-
-
-print("Checking insert insert user provisionally into test user db...")
-assert(dao.insertProvisionalUser2Db("test:insertProvisionalUser2Db", "email@test.com", "123abc456def", test_user_db) == True)
-assert(len(dao.select("SELECT * FROM provisionalUser", test_user_db, False)) == 1)
-assert(dao.select("SELECT * FROM provisionalUser", test_user_db, False)[0][1] == "test:insertProvisionalUser2Db")
+print("Checking insert user provisionally into test user db...")
+assert(dao.insertProvisionalUser2Db("name_test_1", "email@test.com", "123abc456def", test_user_db) == True)
+assert(dao.insertProvisionalUser2Db("name_test_2", "email@test.com", "123abc456def", test_user_db) == True)
+assert(len(dao.select("SELECT * FROM provisionalUser", test_user_db, False)) == 2)
+assert(dao.select("SELECT * FROM provisionalUser", test_user_db, False)[0][0] == 1)
+assert(dao.select("SELECT * FROM provisionalUser", test_user_db, False)[1][0] == 2)
+assert(dao.select("SELECT * FROM provisionalUser", test_user_db, False)[0][1] == "name_test_1")
+assert(dao.select("SELECT * FROM provisionalUser", test_user_db, False)[1][1] == "name_test_2")
 assert(dao.deleteAll("provisionalUser", test_user_db) == True)
 assert(len(dao.select("SELECT * FROM provisionalUser", test_user_db, False)) == 0)
 
+print("Checking delete user provisionally from test user db...")
+assert(dao.insertProvisionalUser2Db("delete_test_1", "delete@test.com", "123abc456def", test_user_db) == True)
+assert(len(dao.select("SELECT * FROM provisionalUser WHERE name={} AND email={}".format("'delete_test_1'", "'delete@test.com'"), test_user_db, False)) == 1)
+assert(dao.deleteProvisionalUser2Db("delete_test_1", "delete@test.com", test_user_db) == True)
+assert(len(dao.select("SELECT * FROM provisionalUser WHERE name={} AND email={}".format("'delete_test_1'", "'delete@test.com'"), test_user_db, False)) == 0)
+
+print("Checking insert user into test user db...")
+assert(dao.insertUser2Db("name_test_1", "email@test.com", "123abc456def", test_user_db) == True)
+assert(len(dao.select("SELECT * FROM user", test_user_db, False)) == 1)
+assert(dao.select("SELECT * FROM user WHERE name={} AND email={}".format("'name_test_1'","'email@test.com'"), test_user_db, False)[0][1] == "name_test_1")
+assert(dao.select("SELECT * FROM user WHERE name={} AND email={}".format("'name_test_1'","'email@test.com'"), test_user_db, False)[0][2] == "email@test.com")
+assert(dao.insertUser2Db("name_test_2", "email2@test.com", "789abc012def", test_user_db) == True)
+assert(len(dao.select("SELECT * FROM user", test_user_db, False)) == 2)
+
+print("Checking delete user from test user db...")
+assert(dao.deleteUserFromDb("name_test_1", "email@test.com", "123abc456def", test_user_db) == True)
+assert(len(dao.select("SELECT * FROM user", test_user_db, False)) == 1)
+assert(dao.deleteUserFromDb("name_test_2", "email2@test.com", "789abc012def", test_user_db) == True)
+assert(len(dao.select("SELECT * FROM user", test_user_db, False)) == 0)
 
 print("Finish All Test.")
